requirements.txt
flask
pyusb
waitress